add -t for test mode. this limits us to one upstream query (AAAA), and prints no output.
look at QR bit to ensure that the answer we received was a response.

do not treat empty answers (no records of that type) as errors.

exit status is now nonzero for usage errors.

exit status is now nonzero if there is no evidence of a DoH service
at the given URI.

Closes #11